# server/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Bansuri Creations API",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "components": {
    "securitySchemes": {
      "adminSession": {
        "type": "apiKey",
        "in": "cookie",
        "name": "admin_session"
      }
    },
...
  },
  "paths": {
...
    "/api/products": {
      "get": {
        "operationId": "listProducts",
        "tags": [
          "Products"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "category",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "List of active products",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string"
                      },
                      "name": {
                        "type": "string"
                      },
                      "slug": {
                        "type": "string"
                      },
                      "description": {
                        "type": "string"
                      },
                      "categoryId": {
                        "type": "string"
                      },
                      "basePrice": {
                        "type": "number"
                      },
                      "originalPrice": {
                        "type": [
                          "number",
                          "null"
                        ]
                      },
                      "imageUrl": {
                        "type": "string"
                      },
                      "images": {
                        "type": "array",
                        "items": {
                          "type": "string"
                        }
                      },
                      "rating": {
                        "type": "number"
                      },
                      "isActive": {
                        "type": "boolean"
                      },
                      "createdAt": {
                        "type": "string"
                      },
                      "updatedAt": {
                        "type": "string"
                      },
                      "variants": {
                        "type": "array",
                        "items": {
                          "type": "object",
                          "properties": {
                            "id": {
                              "type": "string"
                            },
                            "productId": {
                              "type": "string"
                            },
                            "label": {
                              "type": "string"
                            },
                            "price": {
                              "type": [
                                "number",
                                "null"
                              ]
                            },
                            "stock": {
                              "type": "number"
                            },
                            "sku": {
                              "type": "string"
                            },
                            "createdAt": {
                              "type": "string"
                            },
                            "updatedAt": {
                              "type": "string"
                            }
                          },
                          "required": [
                            "id",
                            "productId",
                            "label",
                            "price",
                            "stock",
                            "sku",
                            "createdAt",
                            "updatedAt"
                          ]
                        }
                      },
                      "category": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "name": {
                            "type": "string"
                          },
                          "slug": {
                            "type": "string"
                          },
                          "description": {
                            "type": "string"
                          },
                          "imageUrl": {
                            "type": "string"
                          },
                          "icon": {
                            "type": "string"
                          },
                          "createdAt": {
                            "type": "string"
                          },
                          "updatedAt": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "id",
                          "name",
                          "slug",
                          "description",
                          "imageUrl",
                          "icon",
                          "createdAt",
                          "updatedAt"
                        ]
                      }
                    },
                    "required": [
                      "id",
                      "name",
                      "slug",
                      "description",
                      "categoryId",
                      "basePrice",
                      "originalPrice",
                      "imageUrl",
                      "images",
                      "rating",
                      "isActive",
                      "createdAt",
                      "updatedAt",
                      "variants",
                      "category"
                    ]
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/products/{slug}": {
      "get": {
        "operationId": "getProduct",
        "tags": [
          "Products"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "slug",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Product detail",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "name": {
                      "type": "string"
                    },
                    "slug": {
                      "type": "string"
                    },
                    "description": {
                      "type": "string"
                    },
                    "categoryId": {
                      "type": "string"
                    },
                    "basePrice": {
                      "type": "number"
                    },
                    "originalPrice": {
                      "type": [
                        "number",
                        "null"
                      ]
                    },
                    "imageUrl": {
                      "type": "string"
                    },
                    "images": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    },
                    "rating": {
                      "type": "number"
                    },
                    "isActive": {
                      "type": "boolean"
                    },
                    "createdAt": {
                      "type": "string"
                    },
                    "updatedAt": {
                      "type": "string"
                    },
                    "variants": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "productId": {
                            "type": "string"
                          },
                          "label": {
                            "type": "string"
                          },
                          "price": {
                            "type": [
                              "number",
                              "null"
                            ]
                          },
                          "stock": {
                            "type": "number"
                          },
                          "sku": {
                            "type": "string"
                          },
                          "createdAt": {
                            "type": "string"
                          },
                          "updatedAt": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "id",
                          "productId",
                          "label",
                          "price",
                          "stock",
                          "sku",
                          "createdAt",
                          "updatedAt"
                        ]
                      }
                    },
                    "category": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "name": {
                          "type": "string"
                        },
                        "slug": {
                          "type": "string"
                        },
                        "description": {
                          "type": "string"
                        },
                        "imageUrl": {
                          "type": "string"
                        },
                        "icon": {
                          "type": "string"
                        },
                        "createdAt": {
                          "type": "string"
                        },
                        "updatedAt": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "id",
                        "name",
                        "slug",
                        "description",
                        "imageUrl",
                        "icon",
                        "createdAt",
                        "updatedAt"
                      ]
                    }
                  },
                  "required": [
                    "id",
                    "name",
                    "slug",
                    "description",
                    "categoryId",
                    "basePrice",
                    "originalPrice",
                    "imageUrl",
                    "images",
                    "rating",
                    "isActive",
                    "createdAt",
                    "updatedAt",
                    "variants",
                    "category"
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Product not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "details": {}
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          }
        }
      }
    },
...
    "/api/admin/products/{id}": {
      "put": {
        "operationId": "adminUpdateProduct",
        "tags": [
          "Admin Products"
        ],
        "security": [
          {
            "adminSession": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1
                  },
                  "description": {
                    "type": "string",
                    "minLength": 1
                  },
                  "categoryId": {
                    "type": "string",
                    "minLength": 1
                  },
                  "basePrice": {
                    "type": "integer",
                    "exclusiveMinimum": 0
                  },
                  "originalPrice": {
                    "type": [
                      "integer",
                      "null"
                    ],
                    "exclusiveMinimum": 0
                  },
                  "imageUrl": {
                    "type": "string",
                    "minLength": 1
                  },
                  "images": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "minLength": 1
                    }
                  },
                  "isActive": {
                    "type": "boolean"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Product updated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "name": {
                      "type": "string"
                    },
                    "slug": {
                      "type": "string"
                    },
                    "description": {
                      "type": "string"
                    },
                    "categoryId": {
                      "type": "string"
                    },
                    "basePrice": {
                      "type": "number"
                    },
                    "originalPrice": {
                      "type": [
                        "number",
                        "null"
                      ]
                    },
                    "imageUrl": {
                      "type": "string"
                    },
                    "images": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    },
                    "rating": {
                      "type": "number"
                    },
                    "isActive": {
                      "type": "boolean"
                    },
                    "createdAt": {
                      "type": "string"
                    },
                    "updatedAt": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "id",
                    "name",
                    "slug",
                    "description",
                    "categoryId",
                    "basePrice",
                    "originalPrice",
                    "imageUrl",
                    "images",
                    "rating",
                    "isActive",
                    "createdAt",
                    "updatedAt"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Invalid payload",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "details": {}
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "details": {}
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/products/{id}/variants/{variantId}": {
      "put": {
        "operationId": "adminUpdateVariant",
        "tags": [
          "Admin Products"
        ],
        "security": [
          {
            "adminSession": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "id",
            "in": "path"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "variantId",
            "in": "path"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "label": {
                    "type": "string",
                    "minLength": 1
                  },
                  "price": {
                    "type": "integer",
                    "exclusiveMinimum": 0
                  },
                  "stock": {
                    "type": "integer",
                    "minimum": 0
                  },
                  "sku": {
                    "type": "string",
                    "minLength": 1
                  },
                  "id": {
                    "type": "string"
                  }
                },
                "required": [
                  "label",
                  "stock",
                  "sku"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Variant updated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "productId": {
                      "type": "string"
                    },
                    "label": {
                      "type": "string"
                    },
                    "price": {
                      "type": [
                        "number",
                        "null"
                      ]
                    },
                    "stock": {
                      "type": "number"
                    },
                    "sku": {
                      "type": "string"
                    },
                    "createdAt": {
                      "type": "string"
                    },
                    "updatedAt": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "id",
                    "productId",
                    "label",
                    "price",
                    "stock",
                    "sku",
                    "createdAt",
                    "updatedAt"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Invalid payload",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "details": {}
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "details": {}
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}